pyproject.toml
[project]
name = "osa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "crewai[tools]>=0.130.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai<0.3.0",
    "notion-client>=2.4.0",
    "openai>=1.88.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
]
